habitnow-flask/requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.0
flask-cors==3.0.10
